+++ Passepartout/App/macOS/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Website guidance in provider account screen.
+
 ### Changed
 
 - Improved debug log appearance.